chore: extract the animation params function
