+++ packages/vchart/src/series/bar/bar.ts
@@ -7,7 +7,6 @@ import type { IMark, IMarkProgressiveConfig } from '../../mark/interface';
 import { MarkTypeEnum } from '../../mark/interface/type';
 import { AttributeLevel } from '../../constant';
 import type { Maybe, Datum, DirectionType } from '../../typings';
-import { mergeSpec } from '../../util/spec/merge-spec';
 import { valueInScaleRange } from '../../util/scale';
 import { getRegionStackGroup } from '../../util/data';
 import { getActualNumValue } from '../../util/space';
@@ -33,6 +32,7 @@ import { addVChartProperty } from '../../data/transforms/add-property';
 import { addDataKey, initKeyMap } from '../../data/transforms/data-key';
 import { registerSampleTransform } from '@visactor/vgrammar-core';
 import type { ILabelSpec } from '../../component';
+import { getGroupAnimationParams } from '../util/utils';
 
 export const DefaultBandWidth = 6; // 默认的bandWidth，避免连续轴没有bandWidth
 const RECT_X = `${PREFIX}_rect_x`;
@@ -481,7 +481,7 @@ export class BarSeries<T extends IBarSeriesSpec = IBarSeriesSpec> extends Cartes
   initAnimation() {
     // 这个数据在这个时候拿不到，因为组件还没创建结束，统计和筛选也还没添加。
     // 而且这个值理论上是动态的，建议 监听 viewDataStatisticsUpdate 消息动态更新
-    const animationParams: IBarAnimationParams = {
+    const barAnimationParams: IBarAnimationParams = {
       yField: this._fieldY[0],
       xField: this._fieldX[0],
       direction: this.direction,
@@ -491,25 +491,13 @@ export class BarSeries<T extends IBarSeriesSpec = IBarSeriesSpec> extends Cartes
           : this._yAxisHelper?.getScale(0).scale(0)
     };
     const appearPreset = (this._spec?.animationAppear as IStateAnimateSpec<BarAppearPreset>)?.preset;
-    // 分组数据的dataIndex应该与x轴顺序一致，而非data[DEFAULT_DATA_INDEX]顺序
-    const dataIndex = (datum: any) => {
-      const indexField = this.direction === 'horizontal' ? this._fieldY[0] : this._fieldX[0];
-      const indexValue = datum?.[indexField];
-      const scale = this.direction === 'horizontal' ? this._scaleY : this._scaleX;
-      const index = (scale?.domain?.() ?? []).indexOf(indexValue);
-      // 不应该出现xIndex === -1 || undefined的情况
-      return index || 0;
-    };
-    const dataCount = () => {
-      const scale = this.direction === 'horizontal' ? this._scaleY : this._scaleX;
-      return (scale?.domain?.() ?? []).length ?? 0;
-    };
+    const animationParams = getGroupAnimationParams(this);
 
     this._barMark.setAnimationConfig(
       animationConfig(
-        Factory.getAnimationInKey('bar')?.(animationParams, appearPreset),
+        Factory.getAnimationInKey('bar')?.(barAnimationParams, appearPreset),
         userAnimationConfig(this._barMarkName, this._spec, this._markAttributeContext),
-        { dataIndex, dataCount }
+        animationParams
       )
     );
   }
